pyproject.toml
[project]
name = "grader-ai"
version = "0.0.0"
description = "Grade assignments with LLMs"
readme = "README.md"
authors = [
    { name = "Zijian Zhang", email = "35801754+futrime@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.8.0",
    "openai>=2.24.0",
    "pandas>=3.0.1",
    "pylatexenc>=2.10",
    "python-dotenv>=1.2.1",
    "rich>=14.3.3",
]

[project.scripts]
grader-ai = "grader_ai.cli:main"
grader-web = "grader_ai.web:main"

[build-system]
requires = ["uv_build>=0.10.4,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.4",
]
